% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstArray-class.R
\docType{class}
\name{fstArraySeed-class}
\alias{fstArraySeed-class}
\alias{fstArraySeed}
\alias{fstArray-class}
\alias{fstMatrix-class}
\alias{fstArray}
\title{fst datasets as array-like objects}
\usage{
fstArraySeed(file, old_format = FALSE)

fstArray(file, old_format = FALSE)
}
\arguments{
\item{file}{The path (as a single character string) to the fst file where
the dataset is located.}

\item{old_format}{use \code{TRUE} to read fst files generated with a fst package
version lower than v0.8.0}
}
\value{
An fstArray object for \code{fstArray()}.

An fstArraySeed object for \code{fstArraySeed()}.
}
\description{
We provide 2 classes for representing an (on-disk) fst dataset
as an array-like object in R:
\itemize{
\item fstArray: A high-level class, fstArray extends
\link[DelayedArray:DelayedArray-class]{DelayedArray::DelayedArray}. All the operations available on
\link[DelayedArray:DelayedArray-class]{DelayedArray::DelayedArray} objects work on fstArray objects.
\item fstArraySeed: A low-level class for pointing to a fst dataset. No
operation can be performed directly on an fstArraySeed object. It needs to
be wrapped in a \link[DelayedArray:DelayedArray-class]{DelayedArray::DelayedArray} or fstArray object first.
A fstArray object is just a fstArraySeed wrapped in a
\link[DelayedArray:DelayedArray-class]{DelayedArray::DelayedArray} object.
}
}
\seealso{
\itemize{
\item \link[DelayedArray:DelayedArray-class]{DelayedArray::DelayedArray} objects.
\item \link[DelayedArray:DelayedArray-utils]{DelayedArray::DelayedArray-utils} for common operations on \link[DelayedArray:DelayedArray-class]{DelayedArray::DelayedArray} objects
\item \link[base:array]{base::array} objects in base R.
}
}
\alias{dim,fstArraySeed-method}
\alias{dimnames,fstArraySeed-method}
\alias{extract_array,fstArraySeed-method}
\alias{DelayedArray,fstArraySeed-method}
