% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstArray-package.R
\docType{package}
\name{fstArray-package}
\alias{fstArray-package}
\title{fstArray: 'fst' backend for DelayedArray objects}
\description{
\strong{fstArray} provides an 'fst' backend for \emph{DelayedArray} objects.
}
