% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstArray-package.R
\docType{package}
\name{fstArray}
\alias{fstArray}
\alias{fstArray-package}
\title{fstArray: 'fst' backend for DelayedArray objects}
\description{
**fstArray** provides an 'fst' backend for *DelayedArray* objects.
}
