% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FstFile.R
\docType{class}
\name{FstFile-class}
\alias{FstFile-class}
\alias{.FstFile}
\alias{FstFile}
\alias{summary,FstFile-method}
\alias{nrow,FstFile-method}
\alias{colnames,FstFile-method}
\alias{ncol,FstFile-method}
\alias{dim,FstFile-method}
\alias{type,FstFile-method}
\alias{import,FstFile,ANY,ANY-method}
\alias{export,ANY,FstFile,ANY-method}
\title{FstFile-class: Import, export and introspection of fst-files}
\usage{
FstFile(resource)

\S4method{summary}{FstFile}(object)

\S4method{nrow}{FstFile}(x)

\S4method{colnames}{FstFile}(x)

\S4method{ncol}{FstFile}(x)

\S4method{dim}{FstFile}(x)

\S4method{type}{FstFile}(x)

\S4method{import}{FstFile,ANY,ANY}(
  con,
  format,
  text,
  columns = NULL,
  from = 1,
  to = NULL,
  asDataFrame = TRUE
)

\S4method{export}{ANY,FstFile,ANY}(object, con, format, compress = 50)
}
\arguments{
\item{resource}{character: path to fst-file}

\item{object}{FstFile-object}

\item{x}{FstFile-object}

\item{con}{The connection from which data is loaded or to which data is
saved. If this is a \code{character} vector, it is assumed to be a file name and
a corresponding file connection is created and then closed after exporting
the object. If it is a \linkS4class{BiocFile} derivative, the data is loaded from
or saved to the underlying resource. If missing, the function will return
the output as a character vector, rather than writing to a connection.}

\item{format}{The format of the output. If missing and \code{con} is a
file name, the format is derived from the file extension. This argument is
unnecessary when \code{con} is a derivative of \linkS4class{BiocFile}.}

\item{text}{If \code{con} is missing, this can be a character vector directly
providing the string data to import.}

\item{columns}{Column names to read. The default is to read all columns.}

\item{from}{Read data starting from this row number.}

\item{to}{Read data up until this row number. The default is to read to the last row of the stored dataset.}

\item{asDataFrame}{logical: Whether to return as a DataFrame (TRUE) or data.frame (FALSE)}

\item{compress}{value in the range 0 to 100, indicating the amount of compression to use.
Lower values mean larger file sizes. The default compression is set to 50.}
}
\value{
FstFile-object
}
\description{
FstFile-class: Import, export and introspection of fst-files
}
\examples{
# Save mtcars as an .fst file with default settings
fst_fname <- tempfile(fileext=".fst")
export(mtcars, fst_fname)

# Change compression
export(mtcars, fst_fname, compress=25)

# Import everything
import(fst_fname)

# Import subset of rows and columns
import(fst_fname,
       columns=c("hp", "vs", "carb"),
       from=5, to=10)

# Get file info without loading the .fst-file
ff <- FstFile(fst_fname)

# Dimension
nrow(ff)
ncol(ff)
dim(ff)

# Column names type
colnames(ff)
type(ff)

# Access to the raw summary from the fst package
summary(ff)
}
