% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstArray-class.R
\docType{class}
\name{fstArraySeed-class}
\alias{fstArraySeed-class}
\alias{fstArraySeed}
\alias{fstArray-class}
\alias{fstMatrix-class}
\alias{fstArray}
\title{fst datasets as array-like objects}
\usage{
fstArraySeed(filepath, old_format = FALSE)

fstArray(filepath, old_format = FALSE)
}
\arguments{
\item{filepath}{The path (as a single character string) to the fst file where
the dataset is located or an \link[fst:fst]{fst::fst_table} instance.}

\item{old_format}{use \code{TRUE} to read fst files generated with a fst package
version lower than v0.8.0.}
}
\value{
An \emph{fstArray} object for \code{fstArray()}.

An \emph{fstArraySeed} object for \code{fstArraySeed()}.
}
\description{
We provide 2 classes for representing an (on-disk) fst dataset
as an array-like object in R:
\itemize{
\item \emph{fstArray}: A high-level class, \emph{fstArray} extends
\link[DelayedArray:DelayedArray-class]{DelayedArray::DelayedArray}. All the operations available on
\link[DelayedArray:DelayedArray-class]{DelayedArray::DelayedArray} objects work on \emph{fstArray} objects.
\item \emph{fstArraySeed}: A low-level class for pointing to a fst dataset. No
operation can be performed directly on an \emph{fstArraySeed} object. It needs to
be wrapped in a \link[DelayedArray:DelayedArray-class]{DelayedArray::DelayedArray} or \emph{fstArray} object first.
An \emph{fstArray} object is just an \emph{fstArraySeed} wrapped in a
\link[DelayedArray:DelayedArray-class]{DelayedArray::DelayedArray} object.
}
}
\examples{
# ----------------------------------------------------------------------
# CONSTRUCTION
# ----------------------------------------------------------------------

# Simulate some data in a data frame and write to disk as an 'fst' file
library(fst)
x <- as.data.frame(replicate(10, runif(10000)))
fst_file <- tempfile(fileext = ".fst")
write_fst(x, fst_file)

# Construct an fstArray from an fst file
fst_array <- fstArray(fst_file)
fst_array
# Construct an fstArray from an fst::fst_table
fst_table <- fst(fst_file)
class(fst_table)
fstArray(fst_table)

# ----------------------------------------------------------------------
# dim/dimnames
# ----------------------------------------------------------------------

dim(fst_array)

dimnames(fst_array)
dimnames(fst_array) <- list(paste0("gene", seq_len(nrow(fst_array))),
                            paste0("S", seq_len(ncol(fst_array))))
fst_array

# ----------------------------------------------------------------------
# SLICING (A.K.A. SUBSETTING)
# ----------------------------------------------------------------------

fst_array2 <- drop(fst_array[31:40, c("S3", "S6")])
fst_array2

dim(fst_array2)
as.array(fst_array2)
stopifnot(identical(dim(as.array(fst_array2)), dim(fst_array2)))
stopifnot(identical(dimnames(as.array(fst_array2)), dimnames(fst_array2)))
# ----------------------------------------------------------------------
# SummarizedExperiment OBJECTS WITH DELAYED ASSAYS
# ----------------------------------------------------------------------

library(SummarizedExperiment)
se <- SummarizedExperiment(fst_array)
se

stopifnot(validObject(se, complete = TRUE))

}
\seealso{
\itemize{
\item \code{\link[fst:write_fst]{fst::write_fst()}} for writing a data frame to disk as an 'fst' file.
\item \link[DelayedArray:DelayedArray-class]{DelayedArray::DelayedArray} objects.
\item \link[DelayedArray:DelayedArray-utils]{DelayedArray::DelayedArray-utils} for common operations on \link[DelayedArray:DelayedArray-class]{DelayedArray::DelayedArray} objects
\item \link[base:array]{base::array} objects in base R.
}
}
\alias{dim,fstArraySeed-method}
\alias{dimnames,fstArraySeed-method}
\alias{extract_array,fstArraySeed-method}
\alias{DelayedArray,fstArraySeed-method}
\alias{path,fstArraySeed-method}
