% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstArray-class.R
\docType{class}
\name{fstArraySeed-class}
\alias{fstArraySeed-class}
\alias{fstArraySeed}
\alias{fstArray-class}
\alias{fstMatrix-class}
\alias{fstArray}
\title{fst datasets as array-like objects}
\usage{
fstArraySeed(file, old_format = FALSE)

fstArray(file, old_format = FALSE)
}
\arguments{
\item{file}{The path (as a single character string) to the fst file where
the dataset is located.}

\item{old_format}{use \code{TRUE} to read fst files generated with a fst package
version lower than v0.8.0}
}
\value{
An fstArray object for \code{fstArray()}.

An fstArraySeed object for \code{fstArraySeed()}.
}
\description{
We provide 2 classes for representing an (on-disk) fst dataset
as an array-like object in R:
\itemize{
\item fstArray: A high-level class, fstArray extends
\link[DelayedArray:DelayedArray-class]{DelayedArray::DelayedArray}. All the operations available on
\link[DelayedArray:DelayedArray-class]{DelayedArray::DelayedArray} objects work on fstArray objects.
\item fstArraySeed: A low-level class for pointing to a fst dataset. No
operation can be performed directly on an fstArraySeed object. It needs to
be wrapped in a \link[DelayedArray:DelayedArray-class]{DelayedArray::DelayedArray} or fstArray object first.
A fstArray object is just a fstArraySeed wrapped in a
\link[DelayedArray:DelayedArray-class]{DelayedArray::DelayedArray} object.
}
}
\examples{
# ----------------------------------------------------------------------
# CONSTRUCTION
# ----------------------------------------------------------------------

# Simulate some data in a data frame and write to disk as an 'fst' file
library(fst)
x <- as.data.frame(replicate(10, runif(10000)))
fst_file <- tempfile(fileext = ".fst")
write_fst(x, fst_file)

# Construct an fstArray
fst_array <- fstArray(fst_file)
fst_array

# ----------------------------------------------------------------------
# dim/dimnames
# ----------------------------------------------------------------------

dim(fst_array)

dimnames(fst_array)
dimnames(fst_array) <- list(paste0("gene", seq_len(nrow(fst_array))),
                            paste0("S", seq_len(ncol(fst_array))))
fst_array

# ----------------------------------------------------------------------
# SLICING (A.K.A. SUBSETTING)
# ----------------------------------------------------------------------

fst_array2 <- drop(fst_array[31:40, c("S3", "S6")])
fst_array2

dim(fst_array2)
as.array(fst_array2)
stopifnot(identical(dim(as.array(fst_array2)), dim(fst_array2)))
stopifnot(identical(dimnames(as.array(fst_array2)), dimnames(fst_array2)))
# ----------------------------------------------------------------------
# SummarizedExperiment OBJECTS WITH DELAYED ASSAYS
# ----------------------------------------------------------------------

library(SummarizedExperiment)
se <- SummarizedExperiment(fst_array)
se

stopifnot(validObject(se, complete = TRUE))

}
\seealso{
\itemize{
\item \link[DelayedArray:DelayedArray-class]{DelayedArray::DelayedArray} objects.
\item \link[DelayedArray:DelayedArray-utils]{DelayedArray::DelayedArray-utils} for common operations on \link[DelayedArray:DelayedArray-class]{DelayedArray::DelayedArray} objects
\item \link[base:array]{base::array} objects in base R.
}
}
\alias{dim,fstArraySeed-method}
\alias{dimnames,fstArraySeed-method}
\alias{extract_array,fstArraySeed-method}
\alias{DelayedArray,fstArraySeed-method}
