% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FstArray-package.R, R/FstArray.R
\docType{package}
\name{FstArray-package}
\alias{FstArray-package}
\alias{dimnames,FstArraySeed-method}
\title{FstArray: Bioconductor-friendly fst-files}
\usage{
\S4method{dimnames}{FstArraySeed}(x)
}
\description{
Adds DelayedArray and BiocIO wrappers for using fst-files in Bioconductor. Includes functions for simple IO (FstFile), on-disk matrices (FstArray) and DataFrames (FstFrame).
}
\keyword{internal}
